fix(nifs): memory safety, dirty schedulers, deterministic terrain PRNG
- generate_chunk/serialize_chunk run as dirty CPU-bound NIFs so multi-ms
  chunk work stops blocking normal schedulers.
- Zero chunk resource before init, handle enif_alloc failures, publish the
  resource only when fully initialized; destructor tolerates partial init.
- Section count accounts for tall grass at heightmap+1 (fixes dropped grass
  when max_height %% 16 == 15); signed-arithmetic fix in stone layering.
- Replace srand/rand with splitmix64 so a given seed produces identical
  terrain on every platform; seed noise tables at NIF load.
- compile.nifs reports proper compiler diagnostics on failure; Makefile
  depends on headers and creates priv/; package files include src/Makefile.

+++ lib/minecraft/nif.ex
@@ -26,7 +26,7 @@ defmodule Minecraft.NIF do
   Note that these must be chunk coordinates, as they get multiplied by 16
   in the NIF.
   """
-  @spec generate_chunk(float, float) :: {:ok, any} | {:error, any}
+  @spec generate_chunk(integer, integer) :: {:ok, any} | {:error, any}
   def generate_chunk(_chunk_x, _chunk_z) do
     # Don't raise here, or Dialyzer complains
     :erlang.nif_error("NIF generate_chunk/2 not implemented")
